Moving deployment of logging to conftest.py
